refactor(calculator) : Modify the package names
