+++ README.md
@@ -42,7 +42,7 @@
 
 ---
 # 구현절차
-1. step1 클래스
+1. calculator 클래스
     0. 문자열 입력값을 받는 메소드
     1. 모든 숫자값을 합하는 메소드
     2. 커스텀 구분자가 있는지 확인하는 메소드
