+++ src/main/kotlin/calculator/Calculator.kt
@@ -1,4 +1,4 @@
-package step1
+package calculator
 
 import java.util.regex.Pattern
