+++ src/test/kotlin/calculator/CalculatorTest.kt
@@ -1,4 +1,4 @@
-package step1
+package calculator
 
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatThrownBy
